backend/requirements.txt
# API server
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Data processing
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1

# Embedding models
sentence-transformers==2.2.2
# Specific versions to ensure compatibility
transformers==4.30.2
huggingface-hub==0.23.0
accelerate==0.20.3
torch==2.0.1

# Database
sqlalchemy==2.0.21

# Utilities
python-dotenv==1.0.0
uuid==1.30
requests==2.31.0
tqdm==4.66.1
